# openapi_server/models/usuario_registro.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class UsuarioRegistro(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, nombre_usuario=None, correo_electronico=None, contrasea=None):  # noqa: E501
        """UsuarioRegistro - a model defined in OpenAPI

        :param nombre_usuario: The nombre_usuario of this UsuarioRegistro.  # noqa: E501
        :type nombre_usuario: str
        :param correo_electronico: The correo_electronico of this UsuarioRegistro.  # noqa: E501
        :type correo_electronico: str
        :param contrasea: The contrasea of this UsuarioRegistro.  # noqa: E501
        :type contrasea: str
        """
        self.openapi_types = {
            'nombre_usuario': str,
            'correo_electronico': str,
            'contrasea': str
        }

        self.attribute_map = {
            'nombre_usuario': 'nombreUsuario',
            'correo_electronico': 'correoElectronico',
            'contrasea': 'contraseña'
        }

        self._nombre_usuario = nombre_usuario
        self._correo_electronico = correo_electronico
        self._contrasea = contrasea

    @classmethod
    def from_dict(cls, dikt) -> 'UsuarioRegistro':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UsuarioRegistro of this UsuarioRegistro.  # noqa: E501
        :rtype: UsuarioRegistro
        """
        return util.deserialize_model(dikt, cls)

    @property
    def nombre_usuario(self) -> str:
        """Gets the nombre_usuario of this UsuarioRegistro.


        :return: The nombre_usuario of this UsuarioRegistro.
        :rtype: str
        """
        return self._nombre_usuario

    @nombre_usuario.setter
    def nombre_usuario(self, nombre_usuario: str):
        """Sets the nombre_usuario of this UsuarioRegistro.


        :param nombre_usuario: The nombre_usuario of this UsuarioRegistro.
        :type nombre_usuario: str
        """

        self._nombre_usuario = nombre_usuario

    @property
    def correo_electronico(self) -> str:
        """Gets the correo_electronico of this UsuarioRegistro.


        :return: The correo_electronico of this UsuarioRegistro.
        :rtype: str
        """
        return self._correo_electronico

    @correo_electronico.setter
    def correo_electronico(self, correo_electronico: str):
        """Sets the correo_electronico of this UsuarioRegistro.


        :param correo_electronico: The correo_electronico of this UsuarioRegistro.
        :type correo_electronico: str
        """

        self._correo_electronico = correo_electronico

    @property
    def contrasea(self) -> str:
        """Gets the contrasea of this UsuarioRegistro.


        :return: The contrasea of this UsuarioRegistro.
        :rtype: str
        """
        return self._contrasea

    @contrasea.setter
    def contrasea(self, contrasea: str):
        """Sets the contrasea of this UsuarioRegistro.


        :param contrasea: The contrasea of this UsuarioRegistro.
        :type contrasea: str
        """

        self._contrasea = contrasea